GAN/02_DCGAN_AnimeFaces/requirements.txt
numpy
torch
torchvision
matplotlib
logging
pathlib
pyyaml
Pillow